backend/requirements.txt
django>=5.1,<6
djangorestframework>=3.15,<4
django-cors-headers>=4.6,<5
gunicorn>=23.0,<24
psycopg[binary]>=3.1,<4
